Started work
Nothing is working at the moment but i started the framework for the simulation. Still need to implement all of the functionalities. There are already functions and classes for a smoothed particle hydrodynamics simulation and not just general simulation, even though it can be done. All of glfw need to be remade to run on the gpu.
Note: compatibility with pc without a nvidia gpu can be done and using the current structure there is a base for multithreading