Particles now works
It now finally compiles and works. note to myself: accessing memory on the device is not possible except in kernel or through memcpy to the host.
Anyways i can now store particles in a SoA. Next step is either the grid neighbour search or params. I need both anyways for everything to work