Hard boundary and Active Particle count
The boundaries are now like walls: When a particle would  cross the boundary it gets put onto the boundary and its velocity in that direction gets reversed and multiplied by boundaryDamping from the params. The particle system now has a maxParticle count for its size and numParticles for the actual amount of particles that get simulated.